+++ examples/Resample.hs
@@ -14,10 +14,10 @@ signal ts = map (\(x, y) -> (utcToLocalTime utc x, y)) (TS.toList ts)
 
 main = do
     ts <- TS.loadCSV TS.HasHeader parseISODateTime "testdata/co2.csv"
-    let y3 = 3 * 365 * 24 * 3600
-    let xs = TS.rolling y3 (S.mean . V.fromList) ts
-    toFile def "temp/co2_smooth.svg" $ do
+    let startTime = UTCTime (fromGregorian 1965 1 1) 0
+    let xs = TS.resample startTime (TS.years 1) (S.mean . V.fromList) ts
+    toFile def "temp/co2_resample.svg" $ do
         layout_title .= "CO2 level"
         setColors [opaque blue, opaque red]
         plot (line "CO2" [signal ts])
-        plot (line "Average" [signal xs])
+        plot (line "Yearly average" [signal xs])