+++ CHANGES.md
@@ -4,6 +4,7 @@
   * valueAt has different order of its parameters
   * Added rolling window
   * Added variance and standard deviation
+  * Added resampling operation
 
 
 # Version 0.2 (2016-12-15)